Functions: output_file and output_text
Written function to read un-encrypted output of qr code. Functions: output_file and output_text
Note: We are doing this because qrencode can not process new line "\n", so at the time of qr code generation, we replaced that character with replace_char, and now to get the original text, we have to run one of the following functions described below.